Changes to make everything compile with VS
-Changed a couple of integral types
-Switched ofstream for output in resampling to be text mode
-Fixed Context serialization code
-Fixed integral types in empiricalDistribution. Also fixed ambiguous operator precedence

+++ networkReliabilityCommon/empiricalDistribution.h
@@ -109,7 +109,7 @@ namespace networkReliability
 			{
 				throw std::runtime_error("Loaded data vector had wrong size");
 			}
-			int remainingBitsToRead = bitsRequired - (wholeIntsRequired * sizeof(int)*8);
+			int remainingBitsToRead = (int)(bitsRequired - (wholeIntsRequired * sizeof(int)*8));
 			if(remainingBitsToRead > 0)
 			{
 				ar >> storedBits;
